Phase 1 (4.8.6, part 3): triage the remaining 41 to zero -- ruff check exits clean
Fixed (real issues, not suppressed):
- application.py's bare `self.spotify` (B018, brief's own item 5.2) --
  and this uncovered a real, separate bug while fixing it: SpotifyClient
  .__init__ only STORES its token_source/force_refresh callables, it
  never calls them (confirmed by reading the constructor directly), so
  the old bare access never actually triggered OAuth at all -- contrary
  to its own comment's claim and the wizard's own explicit "genuinely
  long-running (opens the system browser...)" framing of what
  connect_spotify does. Fixed to call auth_manager.get_valid_token()
  directly, which does eagerly perform the real token fetch/OAuth-if-
  needed. Four tests in test_application.py mocked the old `spotify`
  property as a stand-in for "was OAuth triggered" or "neuter it so no
  real browser opens" -- updated all four to mock
  SpotifyAuthManager.get_valid_token instead, including one assertion
  that had to change for a genuine reason: connect_spotify's own reset
  of `_auth_manager = None` is now followed by an immediate real
  reconstruction (accessing `self.auth_manager` to call
  get_valid_token()), so `_auth_manager` is a fresh instance afterward,
  not None as it was under the old (never-actually-triggering) code path.
- sync_service.py's dict-then-index loop -> .items() (PLC0206, 5.3).
- formatting.py's max(int(round(x)), 0) -> max(round(x), 0) (RUF046,
  5.3: round() on a float already returns int).
- main_window.py: unused loop var `plan` -> `_plan` (B007, 5.3); a bare
  `raise ValueError(...)` inside except -> `from error` (B904, 5.3); a
  zip() needing strict=True, confirmed both lists are docstring-
  guaranteed same length (B905, 5.3).
- spotify/client.py's Union[str, Callable[[], str]] -> str | Callable
  (UP007, this is a 3.13-only project).
- library/duplicate_service.py's nested if -> one combined condition
  (SIM102).
- A handful of ordinary test-only fixes: dict-then-slice -> next()
  (RUF015 x4), list concat -> unpacking (RUF005), zip -> itertools.
  pairwise (RUF007, the two lists are intentionally different lengths
  so strict= was never the right fix here), assert False -> raise
  AssertionError (B011 x2), unused unpacked variables underscore-
  prefixed (RUF059 x3), a shadowed builtin renamed in two test helpers
  with no real callers using it as a keyword (A002 x2), open() ->
  Path.open() (PTH123), an unused F841 var either dropped or
  underscore-prefixed with a reason (x2), a literal `.` in a
  pytest.raises match= pattern properly escaped and a genuine `.*`
  marked raw (RUF043 x2), a zip() given strict=True after confirming
  both lists are always exactly 3 items long (B905, _stress_step4).

Suppressed (checked individually, false positives or deliberately
untouched code, never blanket):
- ui/workers.py's try/except/pass in _emit_or_drop -- behaviorally
  identical to contextlib.suppress either way, but CLAUDE.md's own
  item 32/41 standing rule is that this file's control flow is not
  touched even cosmetically, given its segfault history (SIM105).
- audio_fingerprint.py's module-level lazy-singleton `global` pair --
  the correct, idiomatic shape for this pattern, not worth a class
  wrapper (PLW0603).
- matching.py's TRAILING_SEPARATOR_RE and notice.py's own comment --
  both deliberately contain the "ambiguous" character being discussed/
  matched, not typos (RUF001/RUF003).
- callback_server.py's log_message(self, format, ...) -- overrides
  BaseHTTPRequestHandler's own parameter name exactly (A002).
- spotify/client.py's local datetime.now() -- brief's own confirmed
  §5.5 false positive, feeds a human-readable LOCAL wall-clock message,
  converting to UTC would make it wrong for non-UTC users (DTZ005).
- Two bare property accesses inside `pytest.raises(...)` blocks where
  the access itself, not its return value, is the assertion (B018 x3).
- Two S311 (non-cryptographic random) in test-data-generation-only
  scripts, suppressed via per-file-ignore with the real reason.
- Two closures over an outer-loop variable in test_stress_e2e.py,
  confirmed synchronous/immediate-use-only so the loop can't advance
  before the closure is called (B023) -- an assert-blind-Exception
  where the test's own point is "this fails somehow, not which
  exception" (B017).

Verified with a fresh cache throughout, matching this phase's own
established discipline. `uv run ruff check src tests` -> **0 findings,
exits 0** -- the actual target of §4.8. All 24 currently-touched files
re-checked with py_compile. uv run pytest -> 1098 passed, 1 skipped in
82.43s. uv run mypy --strict src/ -> Success, no issues found in 85
source files.

+++ src/seeker/ui/formatting.py
@@ -70,7 +70,7 @@ def format_duration_seconds(seconds: float) -> str:
     negative — a caller passing a stale/negative remaining-time value
     gets "0s" rather than a confusing negative duration.
     """
-    total_seconds = max(int(round(seconds)), 0)
+    total_seconds = max(round(seconds), 0)
 
     if total_seconds < 60:
         return f"{total_seconds}s"
